beckend/requirements.txt
selenium>=4.18
webdriver-manager
schedule>=1.2.0
psutil>=5.9.0
python-dotenv>=1.0.1
backports.zoneinfo>=0.2.1;python_version<"3.9"

fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0

python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
argon2-cffi>=21.3.0
python-dotenv>=1.0.1

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
asyncpg>=0.29.0

# Scraper dependencies
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
